조회 기능 태란 12:40

+++ src/main/java/kr/watchdogs/entity/WD_Member.java
@@ -15,11 +15,11 @@ public class WD_Member {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long idx; // PK, 자동증가컬럼
+	private int idx; // PK, 자동증가컬럼
 	private String user_id;
 	private String user_pw;
 	private String user_name;
 	@Column(columnDefinition = "int default 0")
-	private Long cnt;
+	private int cnt;
 
 }
